iteration with 2 events
