+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/datasource.ts
@@ -18,6 +18,7 @@ import AzureMonitorDatasource from './azure_monitor/azure_monitor_datasource';
 import AzureResourceGraphDatasource from './azure_resource_graph/azure_resource_graph_datasource';
 import ResourcePickerData from './resourcePicker/resourcePickerData';
 import { AzureDataSourceJsonData, AzureMonitorQuery, AzureQueryType, DatasourceValidationResult } from './types';
+import { logAzureMonitorEvent } from './utils/logging';
 import migrateAnnotation from './utils/migrateAnnotation';
 import { datasourceMigrations } from './utils/migrateQuery';
 import { VariableSupport } from './variables';
@@ -66,54 +67,18 @@ export default class Datasource extends DataSourceApi<AzureMonitorQuery, AzureDa
   }
 
   query(options: DataQueryRequest<AzureMonitorQuery>): Observable<DataQueryResponse> {
-    console.log('Azure Monitor query');
-    console.log(options.targets);
     const byType = new Map<AzureQueryType, DataQueryRequest<AzureMonitorQuery>>();
 
     for (const baseTarget of options.targets) {
       // Migrate old query structures
       const target = datasourceMigrations(baseTarget);
 
-      console.log(target);
-
-      let typeSpecific;
-      switch (target.queryType) {
-        case AzureQueryType.AzureMonitor:
-          typeSpecific = {
-            dimensions: target.azureMonitor?.dimensionFilters?.length ?? 0,
-            alias: !!target.azureMonitor?.alias,
-            top: !!target.azureMonitor?.top,
-          };
-          break;
-        case AzureQueryType.LogAnalytics:
-          typeSpecific = {
-            complexity: target.azureLogAnalytics?.query?.match(/\|/g)?.length,
-            format: target.azureLogAnalytics?.resultFormat,
-          };
-          break;
-
-        case AzureQueryType.AzureResourceGraph:
-          typeSpecific = {
-            complexity: target.azureResourceGraph?.query?.match(/\|/g)?.length,
-          };
-          break;
-      }
-
-      console.log(typeSpecific);
-
-      const event = {
-        datasource: target.datasource?.type,
-        hide: target.hide,
-        queryType: target.queryType,
-        ...typeSpecific,
-      };
-
-      console.log(event);
-
       // Skip hidden or invalid queries or ones without properties
       if (!target.queryType || target.hide || !hasQueryForType(target)) {
         continue;
       }
+
+      logAzureMonitorEvent(target, 'grafana_query');
 
       // Initialize the list of queries
       if (!byType.has(target.queryType)) {
